blog/requirements.txt
Django==1.8
Pillow==3.0.0
django-ckeditor==5.0.2
django-disqus==0.5
django-grappelli==2.7.2
django-taggit==0.17.6
djangorestframework==3.3.2
gunicorn==19.4.1
markdown2==2.3.0
mock==1.3.0
pbr==1.8.1
six==1.10.0
sorl-thumbnail==12.3
